+++ components/plugins/autologin.rb
@@ -11,31 +11,37 @@
 #
 # @author Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>
 #
-# @version 0.1.6
+# @version 0.2
 class Arachni::Plugins::AutoLogin < Arachni::Plugin::Base
 
-    MSG_SUCCESS     = 'Form submitted successfully.'
-    MSG_FAILURE     = 'Could not find a form suiting the provided params at: '
-    MSG_NO_RESPONSE = 'Form submitted but no response was returned.'
-    MSG_NO_MATCH    = 'Form submitted but the response did not match the verifier.'
+    STATUSES  = {
+        ok:             'Form submitted successfully.',
+        form_not_found: 'Could not find a form suiting the provided parameters.',
+        check_failed:   'Form submitted but the response did not match the verifier.'
+    }
 
     def prepare
         framework.pause
         print_info 'System paused.'
 
-        @params   = form_parse_request_body( options['params'] )
-        @verifier = Regexp.new( options['check'] )
-
-        @errored = false
+        @parameters = form_parse_request_body( options[:parameters] )
+        @verifier   = Regexp.new( options[:check] )
+        @url        = options[:url].to_s
+        @errored    = false
     end
 
     def run
-        # find the login form
-        login_form = session.find_login_form( url: options['url'], inputs: @params.keys )
+        login_form = session.find_login_form(
+            url:    @url,
+            inputs: @parameters.keys
+        )
 
         if !login_form
-            register_results( code: 0, msg: MSG_FAILURE + options['url'] )
-            print_error MSG_FAILURE + options['url']
+            register_results(
+                status:  :form_not_found,
+                message: STATUSES[:form_not_found]
+            )
+            print_error STATUSES[:form_not_found]
             @errored = true
             return
         end
@@ -43,27 +49,32 @@ class Arachni::Plugins::AutoLogin < Arachni::Plugin::Base
         print_status "Found log-in form with name: #{login_form.name || '<n/a>'}"
 
         # merge the input fields of the form with the user supplied parameters
-        login_form.update @params
+        login_form.update @parameters
 
-        res = login_form.submit( mode: :sync, update_cookies: true, follow_location: false )
-        if !res
-            register_results( code: -1, msg: MSG_NO_RESPONSE )
-            print_error MSG_NO_RESPONSE
-            @errored = true
-            return
-        end
+        response = login_form.submit(
+            mode:            :sync,
+            update_cookies:  true,
+            follow_location: false
+        )
 
-        check_url = res.url
-        body = if res.redirection?
-            check_url = to_absolute( res.headers.location )
-            http.get( check_url, mode: :sync, update_cookies: true, follow_location: true ).body
+        check_url = response.url
+        body = if response.redirection?
+            check_url = to_absolute( response.headers.location )
+            http.get( check_url,
+                mode:            :sync,
+                update_cookies:  true,
+                follow_location: true
+            ).body
         else
-            res.body
+            response.body
         end
 
         if !body.match( @verifier )
-            register_results( code: -2, msg: MSG_NO_MATCH )
-            print_error MSG_NO_MATCH
+            register_results(
+                status:  :check_failed,
+                message: STATUSES[:check_failed]
+            )
+            print_error STATUSES[:check_failed]
             @errored = true
             return
         end
@@ -71,13 +82,19 @@ class Arachni::Plugins::AutoLogin < Arachni::Plugin::Base
         session.login_form = login_form
         session.set_login_check check_url, @verifier
 
-        cookies = http.cookies.inject( {} ){ |h, c| h.merge!( c.simple ) } || {}
+        cookies = http.cookies.inject({}){ |h, c| h.merge!( c.simple ) } || {}
 
-        register_results( code: 1, msg: MSG_SUCCESS, cookies: cookies.dup )
-        print_ok MSG_SUCCESS
+        register_results(
+            status:  :ok,
+            message: STATUSES[:ok],
+            cookies: cookies
+        )
+        print_ok STATUSES[:ok]
 
         print_info 'Cookies set to:'
-        cookies.each_pair { |name, val| print_info( '    * ' + name + ' = ' + val ) }
+        cookies.each do |name, val|
+            print_info "    * #{name.inspect} = #{val.inspect}"
+        end
     end
 
     def clean_up
@@ -93,16 +110,27 @@ class Arachni::Plugins::AutoLogin < Arachni::Plugin::Base
         {
             name:        'AutoLogin',
             description: %q{It looks for the login form in the user provided URL,
-                merges its input fields with the user supplied parameters and sets the cookies
-                of the response and request as framework-wide cookies.
+                merges its input fields with the user supplied parameters and sets
+                the cookies of the response and request as framework-wide cookies.
             },
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>',
-            version:     '0.1.6',
+            version:     '0.2',
             options:     [
-                Options::URL.new( 'url', [true, 'The URL that contains the login form.'] ),
-                Options::String.new( 'params', [true, 'Form parameters to submit. ( username=user&password=pass )'] ),
-                Options::String.new( 'check', [true, 'A pattern which will be used to verify a successful login.
-                    For example, if a logout link only appears when a user is logged in then it can be a perfect choice.'] )
+                Options::String.new( :url,
+                    required:    true,
+                    description: 'The URL that contains the login form.'
+                ),
+                Options::String.new( :parameters,
+                    required:    true,
+                    description: 'Form parameters to submit. ( username=user&password=pass )'
+                ),
+                Options::String.new( :check,
+                    required:    true,
+                    description:
+                        'A pattern which will be used to verify a successful login.
+                        For example, if a logout link only appears when a user is ' +
+                        'logged in then it can be a perfect choice.'
+                )
             ],
             priority:    0 # run before any other plugin
         }